// archives/tdf-root-ui-legacy/api/types.ts
export type PreferredSlotDTO = {
  startAt: string;
  endAt: string;
  teacherId?: number;
  teacherName?: string;
};

export type TrialTeacherAvailabilityDTO = {
  subjectId: number;
  subjectName?: string;
  teacherId?: number | null;
  teacherName?: string | null;
};

export type TrialSlotDTO = {
  subjectId: number;
  teacherId: number;
  teacherName: string;
  slots: PreferredSlotDTO[];
};
// ...
export type SubjectDTO = {
  subjectId: number;
  name: string;
};